requirements.txt
asgiref==3.11.0
certifi==2025.11.12
crispy-bootstrap4==2025.6
distlib==0.4.0
Django==6.0
django-appconf==1.2.0
django-crispy-forms==2.5
django-imagekit==6.0.0
django-storages==1.14.6
filelock==3.20.0
gunicorn==23.0.0
packaging==25.0
pilkit==3.0
pillow==12.0.0
pipenv==2025.0.4
platformdirs==4.5.0
setuptools==80.9.0
sqlparse==0.5.4
tzdata==2025.2
virtualenv==20.35.4
